+++ cdmlib-persistence/src/main/java/eu/etaxonomy/cdm/persistence/dao/hibernate/description/DescriptionElementDaoImpl.java
@@ -47,13 +47,13 @@ public class DescriptionElementDaoImpl extends AnnotatableDaoImpl<DescriptionEle
     }
 
     @Override
-    public int countMedia(DescriptionElementBase descriptionElement) {
+    public long countMedia(DescriptionElementBase descriptionElement) {
         AuditEvent auditEvent = getAuditEventFromContext();
         if(auditEvent.equals(AuditEvent.CURRENT_VIEW)) {
             Query query = getSession().createQuery("select count(media) from DescriptionElementBase descriptionElement join descriptionElement.media media where descriptionElement = :descriptionElement");
             query.setParameter("descriptionElement", descriptionElement);
 
-            return ((Long)query.uniqueResult()).intValue();
+            return ((Long)query.uniqueResult());
         } else {
             // Horribly inefficient, I know, but hard to do at the moment with envers.
             // FIXME Improve this (by improving envers)
